% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center-deprecated}
\alias{center-deprecated}
\title{Matrix centering}
\usage{
center(A)
}
\arguments{
\item{A}{matrix}
}
\value{
\code{A} but row centered
}
\description{
C routine to row-center a matrix
}
\seealso{
\code{\link[=lfa-deprecated]{lfa-deprecated()}}
}
\keyword{internal}
