% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfa.R
\name{center}
\alias{center}
\title{Matrix centering}
\usage{
center(A)
}
\arguments{
\item{A}{matrix}
}
\value{
matrix same dimensions \code{A} but row centered
}
\description{
C routine to row-center a matrix
}
\examples{
center(hgdp_subset)
}
