% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfa-defunct.R
\name{lfa-defunct}
\alias{lfa-defunct}
\title{Defunct functions in package \code{lfa}.}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are: \code{read.bed}, \code{read.tped.recode}, \code{model.gof}, \code{center}
}
