% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfa.R
\name{model.gof}
\alias{model.gof}
\title{LFA model goodness of fit}
\usage{
model.gof(X, LF, B)
}
\arguments{
\item{X}{a matrix of SNP genotypes, i.e. an integer matrix of 0's,
1's, and 2's. Sparse matrices of class Matrix are not supported
(yet).}

\item{LF}{matrix of logistic factors}

\item{B}{number of null datasets to generate - \eqn{B=1} is usualy
sufficient. If computational time/power allows, a few extra
\eqn{B} could be helpful}
}
\value{
vector of p-values for each SNP.
}
\description{
Compute SNP-by-SNP goodness-of-fit when compared to population 
structure. This can be aggregated to determine genome-wide 
goodness-of-fit for a particular value of \eqn{d}.
}
\details{
This function returns p-values for LFA model goodness of fit based
on a simulated null.
}
\note{
Genotype matrix is expected to be a matrix of integers with
values 0, 1, and 2. Currently no support for missing values. Note
that the coding of the SNPs does not affect the algorithm.
}
\examples{
LF <- lfa(hgdp_subset, 4)
gof_4 <- model.gof(hgdp_subset, LF, 3)
LF <- lfa(hgdp_subset, 10)
gof_10 <- model.gof(hgdp_subset, LF, 3)
hist(gof_4)
hist(gof_10)
}
