% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bed.R
\name{read.bed-deprecated}
\alias{read.bed-deprecated}
\title{File input: .bed}
\usage{
read.bed(bed.prefix)
}
\arguments{
\item{bed.prefix}{Path leading to the bed, bim, and fam files.}
}
\value{
Genotype matrix
}
\description{
Reads in genotypes in .bed format with corresponding bim
and fam files
}
\details{
Use plink with --make-bed
}
\seealso{
\code{\link[=lfa-deprecated]{lfa-deprecated()}}
}
\keyword{internal}
