% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling.R
\name{read.bed}
\alias{read.bed}
\title{File input: .bed}
\usage{
read.bed(bed.prefix)
}
\arguments{
\item{bed.prefix}{Path leading to the bed, bim, and fam files.}
}
\value{
Genotype matrix
}
\description{
Reads in genotypes in .bed format with corresponding bim
and fam files
}
\details{
Use plink with --make-bed
}
\examples{
# assuming you have PLINK format HapMap data from: http://pngu.mgh.harvard.edu/~purcell/plink/res.shtml
# run this in the unpacked folder
x = NULL
\dontrun{x = read.bed("hapmap_r23a")}
}
