% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling.R
\name{read.tped.recode}
\alias{read.tped.recode}
\title{Read .tped}
\usage{
read.tped.recode(tped.filename, buffer.size = 5e+08)
}
\arguments{
\item{tped.filename}{Path to your .tped file after tranposing and recoding.}

\item{buffer.size}{Number of characters to keep in the buffer}
}
\value{
genotype matrix with elements 0, 1, 2, and NA.
}
\description{
Reads a .tped format genotype matrix and returns the R
object needed by \code{\link{lfa}}.
}
\details{
Use --transpose and --recode12 on your plink formatted genotypes
to generate the proper tped file. This is a pretty terrible function
that uses a growing matrix for the genotypes so it is to your 
benefit to have as large a \code{buffer.size} as possible.
}
\examples{
#assuming you have a .tped file in the right directory
x = NULL
\dontrun{x = read.tped.recode("file.tped")}
}
