% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trunc.svd}
\alias{trunc.svd}
\title{Truncated singular value decomposition}
\usage{
\method{trunc}{svd}(A, d, adjust = 3, maxit = 1000, m_b = 3,
  tol = 1e-10, V = NULL, seed = 1234, ltrace = FALSE,
  override = FALSE)
}
\value{
list with singular value decomposition.
}
\description{
Truncated singular value decomposition
}
\details{
Performs singular value decomposition but only returns the first
\code{d} singular vectors/values. The truncated SVD utilizes Lanczos
bidiagonalization. See references.

This function was modified from the package irlba 1.0.1 (?) under
GPL. The
of the \code{crossprod()} calls with the C wrapper to \code{dgemv} is
a dramatic difference in larger datasets. Since the wrapper is
technically not a matrix multiplication function, it seemed wise
to make a copy of the function.
}

