% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/af.R
\name{af}
\alias{af}
\title{Allele frequencies}
\usage{
af(X, LF, safety = FALSE)
}
\arguments{
\item{X}{a matrix of SNP genotypes, i.e. an integer matrix of 0's,
1's, and 2's. Sparse matrices of class Matrix are not supported
(yet).}

\item{LF}{Matrix of logistic factors, with intercept. Pass in the
return value from \code{lfa}!}

\item{safety}{optional boolean to bypass checks on the genotype
matrices, which require a non-trivial amount of computation.}
}
\value{
Matrix of individual-specific allele frequencies.
}
\description{
Compute matrix of individual-specific allele frequencies
}
\details{
Computes the matrix of individual-specific allele
frequencies, which has the same dimensions of the genotype matrix.
Be warned that this function could use a ton of memory, as the
return value is all doubles. It could be wise to pass only a
selection of the SNPs in your genotype matrix to get an idea for
memory usage. Use \code{gc} to check memory usage!
}

