% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/af.R
\name{af_snp}
\alias{af_snp}
\title{Allele frequencies for SNP}
\usage{
af_snp(snp, LF, debug_ind = NULL)
}
\value{
vector of allele frequencies
}
\description{
Computes individual-specific allele frequencies for a
single SNP.
}

