% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfa.R
\name{lfa}
\alias{lfa}
\title{Logistic factor analysis}
\usage{
lfa(X, d, adjustments = NULL, override = FALSE, safety = FALSE)
}
\arguments{
\item{X}{a matrix of SNP genotypes, i.e. an integer matrix of 0's,
1's, and 2's. Sparse matrices of class Matrix are not supported
(yet).}

\item{d}{number of logistic factors, including the intercept}

\item{adjustments}{a matrix of adjustment variables to hold fixed 
during estimation.}

\item{override}{optional boolean to bypass Lanczos bidiagonalization
SVD. Usually not advised unless encountering a bug in the SVD code.}

\item{safety}{optional boolean to bypass checks on the genotype
matrices, which require a non-trivial amount of computation.}
}
\value{
matrix of logistic factors, with the intercept at the end.
}
\description{
Fit a factor model of dimension \eqn{d} for binomial data. Returns 
logistic factors.
}
\details{
This function performs logistic factor analysis on SNP data. As it
stands, we follow the convention where \eqn{d=1} is intercept only,
and for \eqn{d>1} we compute \eqn{d-1} singular vectors and postpend
the intercept.
}
\note{
Genotype matrix is expected to be a matrix of integers with
values 0, 1, and 2. Note
that the coding of the SNPs does not affect the algorithm.
}
\examples{
LF <- lfa(hgdp_subset, 4)
dim(LF)
head(LF)
}
