% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_af.R
\name{pca_af}
\alias{pca_af}
\title{PCA Allele frequencies}
\usage{
pca_af(X, d, override = FALSE)
}
\arguments{
\item{X}{A matrix of SNP genotypes, i.e. an integer matrix of 0's,
1's, 2's and \code{NA}s.
Sparse matrices of class Matrix are not supported (yet).}

\item{d}{Number of logistic factors, including the intercept}

\item{override}{Optional boolean to bypass Lanczos bidiagonalization
SVD. Usually not advised unless encountering a bug in the SVD code.}
}
\value{
Matrix of individual-specific allele frequencies.
}
\description{
Compute matrix of individual-specific allele frequencies
via PCA
}
\details{
This corresponds to algorithm 1 in the paper. Only used for
comparison purposes.
}
\examples{
LF = lfa(hgdp_subset, 4)
allele_freqs_lfa = af(hgdp_subset, LF)
allele_freqs_pca = pca_af(hgdp_subset, 4)
summary(abs(allele_freqs_lfa-allele_freqs_pca))
}
