% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.bed}
\alias{read.bed}
\title{File input: .bed}
\usage{
read.bed(bed.prefix, buffer.size = 1e+07)
}
\arguments{
\item{bed.prefix}{Path leading to the bed, bim, and fam files.}
}
\value{
Genotype matrix
}
\description{
Reads in genotypes in .bed format with corresponding bim
and fam files
}
\details{
Use plink with --make-bed
}

