% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfa.R
\name{sHWE}
\alias{sHWE}
\title{Hardy-Weinberg Equilibrium in structure populations}
\usage{
sHWE(X, LF, B)
}
\arguments{
\item{X}{a matrix of SNP genotypes, i.e. an integer matrix of 0's,
1's, and 2's. Sparse matrices of class Matrix are not supported
(yet).}

\item{LF}{matrix of logistic factors}

\item{B}{number of null datasets to generate - \eqn{B=1} is usualy
sufficient. If computational time/power allows, a few extra
\eqn{B} could be helpful}
}
\value{
a vector of p-values for each SNP.
}
\description{
Compute structural Hardy-Weinberg Equilibrium (sHWE) p-values
on a SNP-by-SNP basis. These p-values can be aggregated to 
determine genome-wide goodness-of-fit for a particular value
of \eqn{d}. See \url{https://doi.org/10.1101/240804} for more
details.
}
\examples{
LF <- lfa(hgdp_subset, 4)
gof_4 <- sHWE(hgdp_subset, LF, 3)
LF <- lfa(hgdp_subset, 10)
gof_10 <- sHWE(hgdp_subset, LF, 3)
hist(gof_4)
hist(gof_10)
}
